@@ -16,8 +16,8 @@
     "jabber": "cwuerker@ceusmedia.de"
   },
   "require": {
-    "php": "^7.3 || ^8",
-    "ceus-media/common": "^0.8 || ^0.9 || dev-master"
+    "php": "^8.1",
+    "ceus-media/common": "^0.9 | 0.9.x-dev | ^1.0 | 1.0.x-dev"
   },
   "autoload": {
     "psr-4": {
@@ -25,7 +25,64 @@
     }
   },
   "require-dev": {
+    "ceus-media/doc-creator": "^1.0 | 1.0.x-dev",
     "phpstan/phpstan": "*",
-    "phpstan/phpstan-strict-rules": "*"
+    "phpstan/phpstan-strict-rules": "*",
+    "phpunit/phpunit": "^10.1",
+    "php-parallel-lint/php-parallel-lint": "^1.3 | dev-master",
+    "rector/rector": "*"
+  },
+  "scripts": {
+    "dev:all": [
+      "@dev:test-syntax",
+      "@dev:phpstan",
+      "@dev:phpunit"
+    ],
+    "dev:install": "composer install --quiet",
+    "dev:update": "composer update",
+    "dev:create-docs": [
+      "@dev:install",
+      "php vendor/ceus-media/doc-creator/doc-creator.php --config-file=util/doc.xml"
+    ],
+    "dev:phpunit": [
+      "@dev:install",
+      "XDEBUG_MODE=coverage php vendor/bin/phpunit --configuration util/phpunit.xml"
+    ],
+    "dev:phpstan": [
+      "@dev:install",
+      "XDEBUG_MODE=off vendor/bin/phpstan analyze --configuration=util/phpstan.neon --memory-limit=1G --verbose"
+    ],
+    "dev:phpstan-clear": [
+      "@dev:install",
+      "XDEBUG_MODE=off vendor/bin/phpstan clear-result-cache --verbose"
+    ],
+    "dev:phpstan-save": [
+      "@dev:install",
+      "XDEBUG_MODE=off vendor/bin/phpstan analyse --configuration util/phpstan.neon --generate-baseline util/phpstan-baseline.neon --verbose"
+    ],
+    "dev:rector": [
+      "@dev:install",
+      "XDEBUG_MODE=off vendor/bin/rector process --config=util/rector.php --dry-run"
+    ],
+    "dev:rector-fix": [
+      "@dev:install",
+      "XDEBUG_MODE=off vendor/bin/rector process --config=util/rector.php"
+    ],
+    "dev:test-syntax": [
+      "@dev:install",
+      "vendor/bin/parallel-lint . -j 10 --colors --exclude vendor"
+    ]
+  },
+  "scripts-descriptions": {
+    "dev:all": "Install, test syntax, run PHPStan and PHPUnit",
+    "dev:install": "Installs project dependencies for Development stage",
+    "dev:update": "Updates project dependencies for Development stage",
+    "dev:phpstan": "Validate codebase with PHPStan",
+    "dev:phpstan-clear": "Clear PHPStan cache",
+    "dev:phpstan-save": "Store current results in ignore list",
+    "dev:phpunit": "Run unit tests",
+    "dev:rector": "Shows suggestions of configured Rector rules",
+    "dev:rector-fix": "Apply currently configured Rector rules to codebase",
+    "dev:test-syntax": "Validate code base syntax using PHP linter"
   }
 }
